backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
tavily-python>=0.5.0
httpx>=0.27.0
pyyaml>=6.0.0
mcp>=1.0.0
greenlet>=3.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
